Note there are three public variables that don't pass our
checkstyle.  I'm proposing they be merged as is and then
fixed in subsequent changes.

* refresh:
  Continue to other URIs if one in the list is down. SPY-60.
  Revert "SPY-37 & SPY-38: Fixed redistribution performance issue"
  No need for old debugging string in test.
  Add compatibility with netty 3.2.0+.
  Refactored tap message classes.
  SPY-37 & SPY-38: Fixed redistribution performance issue
  Made vbmap in MultiKey operation synchronized
  Don't reconnect when a tap connection finishes.
  Flush the PrintWriter in TapMessagePrinter
  Removed extra variables in tapCustom header
  Removed a print line statement from TestConfig

+++ src/main/java/net/spy/memcached/tapmessage/MessageBuilder.java
@@ -43,7 +43,7 @@ public class MessageBuilder {
     message.setFlags(TapFlag.DUMP);
   }
 
-  public void specifyVbuckets(int[] vbucketlist) {
+  public void specifyVbuckets(short[] vbucketlist) {
     message.setVbucketlist(vbucketlist);
     message.setFlags(TapFlag.LIST_VBUCKETS);
   }
@@ -56,7 +56,7 @@ public class MessageBuilder {
     message.setFlags(TapFlag.KEYS_ONLY);
   }
 
-  public void takeoverVbuckets(int[] vbucketlist) {
+  public void takeoverVbuckets(short[] vbucketlist) {
     message.setVbucketlist(vbucketlist);
     message.setFlags(TapFlag.TAKEOVER_VBUCKETS);
   }
